PERA-3219 Add pingInterval for WC socket and update session cache
